A Decibel to Gain (Power/Voltage) convertor
This code opens a GUI that allows the user to enter a value in Decibels and convert it to Power or Voltage Gain. It also allows the user to enter the Power or Voltage Gain value and convert it to respective decibels. 
The GUI also has a button that opens the following web link, (used for demonstration purposes only)
https://ece.northeastern.edu/courses/eece2150/2017fa/supplementary/db.pdf